Bug fixing, redisson client configuration

+++ src/main/java/com/sport/support/employee/application/port/in/AddEmployeeCommand.java
@@ -2,7 +2,7 @@ package com.sport.support.employee.application.port.in;
 
 import com.sport.support.employee.adapter.in.web.payload.AddEmployeeRequest;
 import com.sport.support.employee.adapter.out.persistence.enumeration.EmployeeType;
-import com.sport.support.infrastructure.common.Money;
+import com.sport.support.infrastructure.common.money.Money;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
@@ -19,8 +19,8 @@ public class AddEmployeeCommand {
    public AddEmployeeCommand(AddEmployeeRequest request) {
       this.username = request.getUsername();
       this.type = request.getType();
-      this.baseSalary = new Money(request.getBaseSalary());
-      this.bonus = new Money(request.getBonus());
+      this.baseSalary = new Money(request.getBaseSalary().doubleValue());
+      this.bonus = new Money(request.getBonus().doubleValue());
       this.startDate = request.getStartDate();
    }
 }
